refactor(): remove redundant imported modules - move Auth util file to utils directory [Delivers #156569762]
